refactor: move utils functions to service class
